Edit method parameters combined article.
Update samples, provide better definition for passing
reference parameters and value parameters by reference
or by value.
