--- samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideParameters/CS/Project.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworks>net6.0</TargetFrameworks>
-  </PropertyGroup>   
-</Project>
